+++ NEWS.md
@@ -6,6 +6,7 @@
 - Support Typst version 0.12
 - Throw a `ContextError` for context values of an incorrect type
 - The `preamble` used in `render` and some `show` methods can now be specified using `set_preamble`
+- `render` now supports the `ignorestatus` keyword parameter
 
 ### Bug Fixes